#17 Refactor user inputs into separate component file See: https://github.com/df-map-archive/dfma-html5-map-viewer/issues/17
